smart-inbox-cleaner/requirements.txt
streamlit==1.35.0
imapclient==3.0.0
python-dotenv==1.0.1
pandas==2.2.2